Sppliting work into 1d and 2d cases to fix issue
First I tried running the simulation directly in the 2D case with
certain simplifications like constant bathymetry. However, even when the
source term is not being integrated in time, the Temperature is not
being transported as a passive scalar as expected.

I will then start with the 1D case in order to fix the behaviour of the
temperature before going back to 2D.